app/config.py
import os
from functools import lru_cache

from dotenv import load_dotenv
from pydantic import Field
from pydantic_settings import BaseSettings, SettingsConfigDict

# Drop EMPTY inherited shell vars before loading .env so they don't shadow
# real values from the dotfile (common Langfuse-keys footgun). Non-empty
# shell vars are kept as-is — that lets us override DATABASE_URL/REDIS_URL
# from the command line without touching .env.
for _key in list(os.environ):
    if not os.environ[_key]:
        del os.environ[_key]
load_dotenv(override=False)


class Settings(BaseSettings):
    model_config = SettingsConfigDict(
        env_file=".env",
        env_file_encoding="utf-8",
        extra="ignore",
    )

    # LLM
    openrouter_api_key: str = Field(default="")
    openrouter_base_url: str = "https://openrouter.ai/api/v1"
    openrouter_http_referer: str = ""
    openrouter_x_title: str = "RAG API homework"

    # Storage
    database_url: str = "postgresql://rag:rag@localhost:5432/rag"
    redis_url: str = "redis://localhost:6379/0"

    # Embeddings
    embedding_model: str = "sentence-transformers/all-MiniLM-L6-v2"
    embedding_dim: int = 384

    # RAG
    rag_top_k: int = 3
    chunk_size_tokens: int = 500
    chunk_overlap_tokens: int = 50

    # Cache
    # ТЗ §5 prescribes 0.92, but that threshold is calibrated for OpenAI
    # embeddings (which cluster tighter near 1.0). MiniLM-L6-v2 produces
    # similarities ~0.7-0.85 for paraphrases of the same question, so we
    # use a more permissive threshold and document the trade-off in the
    # report. Override via CACHE_SIMILARITY_THRESHOLD env var if you swap
    # the embedding model.
    cache_similarity_threshold: float = 0.80
    cache_ttl_seconds: int = 3600

    # Concurrency / timeouts
    max_concurrent_llm_calls: int = 20
    llm_request_timeout_seconds: int = 15

    # Langfuse
    langfuse_public_key: str = ""
    langfuse_secret_key: str = ""
    langfuse_host: str = "https://cloud.langfuse.com"


@lru_cache(maxsize=1)
def get_settings() -> Settings:
    return Settings()
