pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "scripts"]

[project]
name = "rag-api"
version = "0.1.0"
description = "Production-ready RAG API with streaming, semantic cache, rate limit, multi-provider fallback"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "openai>=1.55.0",
    "sentence-transformers>=3.3.0",
    "psycopg[binary]>=3.2.0",
    "psycopg-pool>=3.2.0",
    "pgvector>=0.3.0",
    "redis>=5.2.0",
    "langchain-text-splitters>=0.3.0",
    "tiktoken>=0.8.0",
    "pypdf>=5.1.0",
    "httpx>=0.28.0",
    "sse-starlette>=2.1.0",
    "python-dotenv>=1.0.0",
    "langfuse>=2.55.0",
    "numpy>=1.26.0,<2.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
